+++ BUILD.gn
@@ -1139,7 +1139,6 @@ config("grpc_config") {
         "include/grpc++/support/sync_stream.h",
         "include/grpc++/support/time.h",
         "include/grpcpp/alarm.h",
-        "include/grpcpp/alarm_impl.h",
         "include/grpcpp/channel.h",
         "include/grpcpp/client_context.h",
         "include/grpcpp/completion_queue.h",
@@ -1209,10 +1208,8 @@ config("grpc_config") {
         "include/grpcpp/impl/rpc_service_method.h",
         "include/grpcpp/impl/serialization_traits.h",
         "include/grpcpp/impl/server_builder_option.h",
-        "include/grpcpp/impl/server_builder_option_impl.h",
         "include/grpcpp/impl/server_builder_plugin.h",
         "include/grpcpp/impl/server_initializer.h",
-        "include/grpcpp/impl/server_initializer_impl.h",
         "include/grpcpp/impl/service_type.h",
         "include/grpcpp/resource_quota.h",
         "include/grpcpp/security/auth_context.h",
