+++ gRPC-C++.podspec
@@ -77,7 +77,6 @@ Pod::Spec.new do |s|
     ss.header_mappings_dir = 'include/grpcpp'
 
     ss.source_files = 'include/grpcpp/alarm.h',
-                      'include/grpcpp/alarm_impl.h',
                       'include/grpcpp/channel.h',
                       'include/grpcpp/client_context.h',
                       'include/grpcpp/completion_queue.h',
@@ -143,10 +142,8 @@ Pod::Spec.new do |s|
                       'include/grpcpp/impl/rpc_service_method.h',
                       'include/grpcpp/impl/serialization_traits.h',
                       'include/grpcpp/impl/server_builder_option.h',
-                      'include/grpcpp/impl/server_builder_option_impl.h',
                       'include/grpcpp/impl/server_builder_plugin.h',
                       'include/grpcpp/impl/server_initializer.h',
-                      'include/grpcpp/impl/server_initializer_impl.h',
                       'include/grpcpp/impl/service_type.h',
                       'include/grpcpp/resource_quota.h',
                       'include/grpcpp/security/auth_context.h',
