+++ include/grpcpp/impl/server_builder_plugin.h
@@ -24,13 +24,10 @@
 #include <grpcpp/support/channel_arguments.h>
 #include <grpcpp/support/config.h>
 
-namespace grpc_impl {
-
-class ServerInitializer;
-}  // namespace grpc_impl
 namespace grpc {
 
 class ServerBuilder;
+class ServerInitializer;
 
 /// This interface is meant for internal usage only. Implementations of this
 /// interface should add themselves to a \a ServerBuilder instance through the
@@ -47,10 +44,10 @@ class ServerBuilderPlugin {
 
   /// InitServer will be called in ServerBuilder::BuildAndStart(), after the
   /// Server instance is created.
-  virtual void InitServer(grpc_impl::ServerInitializer* si) = 0;
+  virtual void InitServer(ServerInitializer* si) = 0;
 
   /// Finish will be called at the end of ServerBuilder::BuildAndStart().
-  virtual void Finish(grpc_impl::ServerInitializer* si) = 0;
+  virtual void Finish(ServerInitializer* si) = 0;
 
   /// ChangeArguments is an interface that can be used in
   /// ServerBuilderOption::UpdatePlugins
