--- include/grpcpp/security/cronet_credentials_impl.h
@@ -1,33 +1,0 @@
-/*
- *
- * Copyright 2019 gRPC authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-#ifndef GRPCPP_SECURITY_CRONET_CREDENTIALS_IMPL_H
-#define GRPCPP_SECURITY_CRONET_CREDENTIALS_IMPL_H
-
-#include <memory>
-
-namespace grpc_impl {
-
-class ChannelCredentials;
-
-/// Credentials for a channel using Cronet.
-std::shared_ptr<ChannelCredentials> CronetChannelCredentials(void* engine);
-
-}  // namespace grpc_impl
-
-#endif  // GRPCPP_SECURITY_CRONET_CREDENTIALS_IMPL_H
