+++ src/cpp/server/load_reporter/load_reporting_service_server_builder_plugin.h
@@ -39,10 +39,10 @@ class LoadReportingServiceServerBuilderPlugin : public ServerBuilderPlugin {
   void UpdateServerBuilder(ServerBuilder* builder) override;
 
   // Registers the load reporter service.
-  void InitServer(grpc_impl::ServerInitializer* si) override;
+  void InitServer(ServerInitializer* si) override;
 
   // Starts the load reporter service.
-  void Finish(grpc_impl::ServerInitializer* si) override;
+  void Finish(ServerInitializer* si) override;
 
   void ChangeArguments(const std::string& name, void* value) override {}
   void UpdateChannelArguments(grpc::ChannelArguments* args) override {}
