+++ src/cpp/server/server_builder.cc
@@ -334,7 +334,7 @@ std::unique_ptr<grpc::Server> ServerBuilder::BuildAndStart() {
       std::move(acceptors_), resource_quota_,
       std::move(interceptor_creators_)));
 
-  grpc_impl::ServerInitializer* initializer = server->initializer();
+  ServerInitializer* initializer = server->initializer();
 
   // Register all the completion queues with the server. i.e
   //  1. sync_server_cqs: internal completion queues created IF this is a sync
